SpiderWeb Trap Working
The web slows the player movement speed down. Ideally I want it to pull the player back in when the player is not pressing a button. Beyond that, you should be able to break the connection to the web either with button mashing or shooting it etc.
I also want to scale the web in the direction the player is moving away from it so it stretches.