Made a "Minimum Room" swatch
When in the Rule Tile Palette, with the paint brush selected, select the entire 6x4 room to use as your brush and paint away. This is a lot less tedious than using the single tile, since this is the minimum size a viable room can be. Useful for carving hallways or extending things, but more dynamic than the rectangle paint tool.